Minor changes to match ASGI Specification (#77)
* Change websocket.receive message structure

This ensures that the ``text`` and ``bytes`` keys are always present,
as I think the ASGI specification requires. I base this on,

  Exactly one of ``bytes`` or ``text`` must be non-``None``.

which implies both should be present.

* Ensure ASGI Frameworks receive http.disconnect messages

This ensures the framework receives the disconnect message when the
connection is lost. Which I think is a requirement of the ASGI
specification.